GitButler Integration Commit
This is an integration commit for the virtual branches that GitButler is tracking.

Due to GitButler managing multiple virtual branches, you cannot switch back and
forth between git branches and virtual branches easily. 

If you switch to another branch, GitButler will need to be reinitialized.
If you commit on this branch, GitButler will throw it away.

Here are the branches that are currently applied:
 - Auth (refs/gitbutler/Auth)
   branch head: 32e08d20526f5ae624df2f3279f78d960740e9c3
   - src/app.rs
   - Cargo.lock
   - src/main.rs
   - .gitignore
   - src/routes/signup.rs
   - Cargo.toml
   - src/state.rs
   - src/components/nav.rs
   - src/auth.rs
   - src/components/member.rs
   - src/models/member.rs
   - src/components.rs
   - migrations/20240127205121_init.up.sql
 - Virtual branch (refs/gitbutler/Virtual-branch)
   branch head: 1cd6b49f8926eb92dab27c91856c4b99160e95ed
   - infra/modules/network/vpc.tf
   - infra/modules/network/variables.tf
   - infra/environment.auto.tfvars
   - infra/environment.tf
   - Dockerfile
   - infra/secrets.tf
   - infra/main.tf
   - infra/backend.tf
   - infra/modules/s3Backend/variables.tf
   - infra/modules/s3Backend/s3.tf
   - infra/modules/s3Backend/dynamodb.tf
 - main (refs/gitbutler/main)

Your previous branch was: refs/heads/main

The sha for that commit was: 02ff7a13249875a7c7fad327643842727cce90f0

For more information about what we're doing here, check out our docs:
https://docs.gitbutler.com/features/virtual-branches/integration-branch